requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
asyncpg==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
PyJWT==2.8.0